layout: pull heading block into front matter
Pull the heading.html include into the front matter for card layout and
default (text) layout. You can use "no_page_heading: true" to switch
this off and set a manual heading block.

By default the page heading is the page title. If a different h1 is given
in the front matter, that will override the page title.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>